+++ pkg/validationrule/validation_rule.go
@@ -1,6 +1,7 @@
 package validationrule
 
 import (
+	"fmt"
 	"reflect"
 
 	"github.com/fusakla/promruval/v3/pkg/config"
@@ -38,6 +39,7 @@ func New(name string, scope config.ValidationScope) *ValidationRule {
 type ValidationRule struct {
 	name       string
 	scope      config.ValidationScope
+	onlyIf     []ValidatorWithDetails
 	validators []ValidatorWithDetails
 }
 
@@ -45,8 +47,20 @@ func (r *ValidationRule) Validators() []ValidatorWithDetails {
 	return r.validators
 }
 
+func (r *ValidationRule) OnlyIf() []ValidatorWithDetails {
+	return r.onlyIf
+}
+
 func (r *ValidationRule) AddValidator(newValidator validator.Validator, additionalDetails string) {
 	r.validators = append(r.validators, &validatorWithAdditionalDetails{
+		Validator:         newValidator,
+		additionalDetails: additionalDetails,
+		name:              reflect.TypeOf(newValidator).Elem().Name(),
+	})
+}
+
+func (r *ValidationRule) AddOnlyIfValidator(newValidator validator.Validator, additionalDetails string) {
+	r.onlyIf = append(r.onlyIf, &validatorWithAdditionalDetails{
 		Validator:         newValidator,
 		additionalDetails: additionalDetails,
 		name:              reflect.TypeOf(newValidator).Elem().Name(),
@@ -61,10 +75,26 @@ func (r *ValidationRule) Scope() config.ValidationScope {
 	return r.scope
 }
 
+func validatorTextWithScope(v ValidatorWithDetails, scope config.ValidationScope) string {
+	scopeText := string(scope)
+	if scope == config.AllRulesScope {
+		scopeText = "Rule"
+	}
+	return fmt.Sprintf("%s %s", scopeText, v.String())
+}
+
 func (r *ValidationRule) ValidationTexts() []string {
 	validationTexts := make([]string, 0, len(r.validators))
 	for _, v := range r.validators {
-		validationTexts = append(validationTexts, v.String())
+		validationTexts = append(validationTexts, validatorTextWithScope(v, r.Scope()))
+	}
+	return validationTexts
+}
+
+func (r *ValidationRule) OnlyIfValidationTexts() []string {
+	validationTexts := make([]string, 0, len(r.onlyIf))
+	for _, v := range r.onlyIf {
+		validationTexts = append(validationTexts, validatorTextWithScope(v, validator.Scope(v.Name())))
 	}
 	return validationTexts
 }